Add for_consecutive_items helper
There are a few linters we have in mind where finding groups of
consecutive items in a list and executing some code based on those items
would be useful.

Add the `for_consecutive_items` helper to do most of the heavy-lifting
for these linters. The code was adapted from haml-lint and improved upon
to have a slightly-more-convenient interface.