DRY up RedundantDiv with captures
The `RedundantDiv` linter included duplicate logic for both `class` and
`id` attributes. Introducing the concept of a capture allows us to
define a single matcher that can match both situations, and we can then
differentiate between the two matches in code. This results in DRYer
code.

To support this, add the concept of a matcher to the system, and define
the basic set of matchers needed to support this, along with the DSL
methods that make working with them convenient.