Wrap parse exceptions
We want to explicitly control the control flow of the application due to
parse exceptions. Wrap the `Slim::Parser::SyntaxError` exception with
our own exception (which just inherits from it for now, but this can be
changed later if needed) so we can isolate ourselves from upstream
changes to the exception if we need to.